pyproject.toml
[project]
name = "workraft"
version = "0.5.0"
description = "A simple, lightweight, database-only, worker library in Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiomysql>=0.2.0",
    "asyncio>=3.4.3",
    "beartype>=0.19.0",
    "cryptography>=43.0.1",
    "fire>=0.7.0",
    "loguru>=0.7.2",
    "pydantic>=2.9.2",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = ["pre-commit", "pytest", "mkdocs"]

[tool.ruff]
extend-include = ["*.ipynb"]
src = []

[tool.ruff.lint]
fixable = ["I001", "F401"]
ignore = ["E402", "E721", "E731", "E741", "F722", "F821"]
ignore-init-module-imports = true
select = ["E", "F", "I001"]

[tool.ruff.lint.isort]
combine-as-imports = true
extra-standard-library = ["typing_extensions"]
lines-after-imports = 2
order-by-type = false

[tool.pyright]
include = ["src", "tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["workraft"]
